+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/model/NodeChildData.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -50,6 +50,7 @@ import javax.lang.model.element.ElementKind;
 import javax.lang.model.type.TypeMirror;
 
 import com.oracle.truffle.dsl.processor.ProcessorContext;
+import com.oracle.truffle.dsl.processor.expression.DSLExpression;
 
 public class NodeChildData extends MessageContainer {
 
@@ -75,12 +76,18 @@ public class NodeChildData extends MessageContainer {
     private final Cardinality cardinality;
     private final AnnotationValue executeWithValue;
 
+    private final String implicitCreate;
+    private DSLExpression implicitCreateExpression;
+
+    private final String uncached;
+    private DSLExpression uncachedExpression;
+
     private List<NodeExecutionData> executeWith = Collections.emptyList();
 
     private NodeData childNode;
 
     public NodeChildData(Element sourceElement, AnnotationMirror sourceMirror, String name, TypeMirror nodeType, TypeMirror originalNodeType, Element accessElement, Cardinality cardinality,
-                    AnnotationValue executeWithValue) {
+                    AnnotationValue executeWithValue, String implicitCreate, String uncached) {
         this.sourceElement = sourceElement;
         this.sourceAnnotationMirror = sourceMirror;
         this.name = name;
@@ -89,6 +96,8 @@ public class NodeChildData extends MessageContainer {
         this.accessElement = accessElement;
         this.cardinality = cardinality;
         this.executeWithValue = executeWithValue;
+        this.implicitCreate = implicitCreate;
+        this.uncached = uncached;
     }
 
     public boolean needsGeneratedField() {
@@ -101,6 +110,42 @@ public class NodeChildData extends MessageContainer {
 
     public List<NodeExecutionData> getExecuteWith() {
         return executeWith;
+    }
+
+    public boolean isImplicit() {
+        return implicitCreate != null;
+    }
+
+    public String getImplicitCreate() {
+        return implicitCreate;
+    }
+
+    public DSLExpression getImplicitCreateExpression() {
+        assert isImplicit();
+        return implicitCreateExpression;
+    }
+
+    public void setImplicitCreateExpression(DSLExpression implicitCreateExpression) {
+        assert isImplicit();
+        this.implicitCreateExpression = implicitCreateExpression;
+    }
+
+    public boolean isAllowUncached() {
+        return uncached != null;
+    }
+
+    public String getUncached() {
+        return uncached;
+    }
+
+    public DSLExpression getUncachedExpression() {
+        assert isAllowUncached();
+        return uncachedExpression;
+    }
+
+    public void setUncachedExpression(DSLExpression uncachedExpression) {
+        assert isAllowUncached();
+        this.uncachedExpression = uncachedExpression;
     }
 
     public void setExecuteWith(List<NodeExecutionData> executeWith) {